requirements.txt
python-socketio>=5.3.0
nest_asyncio
python-engineio>=4.2.0
aiohttp>=3.4
six>=1.9.0
websockets>=7.0
websocket-client>=0.54.0
setuptools
requests